refactor: reorganize technology and hero data structures into constant files
