+++ data/locales/en/hero.json
@@ -1,0 +1,4 @@
+{
+   "introduction": "I'm Karol Jaroń",
+   "goalsDescription": "My goal is to write maintainable, clean and understandable code to process development was enjoyable."
+}
